Moved iFrame adapter to new file
